refactor: stand occupation to its own service
