+++ app/Jobs/Stand/OccupyStands.php
@@ -2,7 +2,7 @@
 
 namespace App\Jobs\Stand;
 
-use App\Services\Stand\StandService;
+use App\Services\Stand\StandOccupationService;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
@@ -11,7 +11,7 @@ class OccupyStands implements ShouldQueue
 {
     use Dispatchable, Queueable;
 
-    public function handle(StandService $standService): void
+    public function handle(StandOccupationService $standService): void
     {
         $standService->setOccupiedStands();
     }
